requirements.txt
fastapi==0.111.1
PyJWT==2.8.0
pylint==3.2.6
python-dotenv==1.0.1
SQLAlchemy==2.0.31
bcrypt==4.2.0 
passlib==1.7.4
black==24.4.2
flake8==7.1.0